+++ yash-env/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The methods of `System` and `SystemEx` that return a future now return
   `FlexFuture` instead of `Pin<Box<dyn Future>>`:
     - `System::kill`
+    - `System::raise`
 - External dependency versions:
     - Rust 1.85.0 → 1.86.0